+++ components/Input.vue
@@ -4,7 +4,7 @@
         :class="[sizes[size]]"
         :type="$attrs.type || 'text'"
         @input="$emit('input', $event.target.value)"
-        class="border-0 rounded-lg text-ui-gray-400 disabled:cursor-not-allowed disabled:bg-ui-gray-900 bg-ui-gray-800 hover:bg-ui-gray-900 focus:outline-none focus:bg-ui-gray-900 focus:ring-2 focus:ring-ui-focus"
+        class="border-0 rounded-lg text-ui-gray-400 disabled:cursor-not-allowed disabled:bg-ui-gray-900 bg-ui-gray-800 hover:bg-ui-gray-900 focus:outline-none focus:ring-0"
     />
 </template>